Move NameContainsKeywordsPredicate from model.person to model
Since this class is used by Person and TuitionClass, it makes more sense
to put it under model.

Add test cases for NameContainsKeywordsPredicate
