+++ src/main/java/seedu/address/model/NameContainsKeywordsPredicate.java
@@ -1,9 +1,10 @@
-package seedu.address.model.person;
+package seedu.address.model;
 
 import java.util.List;
 import java.util.function.Predicate;
 
 import seedu.address.commons.util.StringUtil;
+import seedu.address.model.person.Person;
 import seedu.address.model.tuitionclass.TuitionClass;
 
 /**
